+++ api/apps/user-service/src/models/auth/handlers/login.handler.ts
@@ -1,47 +1,11 @@
 import type { RequestHandler } from 'express';
 import prisma from '../../../prismaClient';
 import { comparePassword } from '@libs/helpers/bcrypt';
-import jwt, { type Secret, type SignOptions } from 'jsonwebtoken';
+import { generateAccessToken, generateRefreshToken } from '@libs/helpers/jwt';
 
 type LoginBodyProps = {
   email: string;
   password: string;
-};
-
-type GenerateTokenProps = {
-  userId: string;
-  secret: Secret;
-  exp: SignOptions['expiresIn'];
-};
-
-type GenerateAccessOrRefreshTokenProps = {
-  userId: string;
-};
-
-const accessTokenSecret: Secret = process.env.ACCESS_TOKEN_SECRET!;
-const refreshTokenSecret: Secret = process.env.REFRESH_TOKEN_SECRET!;
-
-const accessTokenExp = process.env.ACCESS_TOKEN_EXP! as SignOptions['expiresIn'];
-const refreshTokenExp = process.env.REFRESH_TOKEN_EXP! as SignOptions['expiresIn'];
-
-const generateToken = ({ userId, secret, exp }: GenerateTokenProps): string => {
-  return jwt.sign({ userId }, secret, { expiresIn: exp });
-};
-
-const generateAccessToken = ({ userId }: GenerateAccessOrRefreshTokenProps) => {
-  return generateToken({
-    userId,
-    secret: accessTokenSecret,
-    exp: accessTokenExp,
-  });
-};
-
-const generateRefreshToken = ({ userId }: GenerateAccessOrRefreshTokenProps) => {
-  return generateToken({
-    userId,
-    secret: refreshTokenSecret,
-    exp: refreshTokenExp,
-  });
 };
 
 const loginHandler: RequestHandler = async (req, res) => {
